requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.34
seaborn==0.13.2
matplotlib==3.9.2
python-dotenv==1.0.1
PyMySQL==1.1.1